Add files via upload
TLauncher may have spyware. I didn't create TLauncher so I can't be too sure. I know people who have had it for years and nothing bad has happened. Since Tlauncher is a third-party software not supported by Mojang, it is very well not within the legal boundaries of the game. Using third-party software to run games certainly has a lot of risk attached to it. But SkaiaCraft is safe :)